requirements.txt
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
azure-functions==1.17.0
numpy==1.24.3
scipy==1.10.1